digital wallet

+++ DigitalWallet/src/main/java/Currency.java
@@ -1,8 +1,10 @@
+import java.math.BigDecimal;
+
 public enum Currency {
-    DOLAR(1.0f),
-    REAL(6.0f);
-    public final float value;
-    Currency(float value) {
+    DOLAR(BigDecimal.valueOf(1)),
+    REAL(BigDecimal.valueOf(6));
+    public final BigDecimal value;
+    Currency(BigDecimal value) {
         this.value = value;
     }
 }
