+++ DigitalWallet/src/main/java/Transaction.java
@@ -1,3 +1,4 @@
+import java.math.BigDecimal;
 import java.time.Instant;
 
-public record Transaction(String ID, Account sourceAccount, Account destinationAccount, float amount, Currency currency, Instant timestamp) { }
+public record Transaction(String ID, Account sourceAccount, Account destinationAccount, BigDecimal amount, Currency currency, Instant timestamp) { }
